Projects/OCR_to_XLSX/requirements.txt
streamlit
google-genai
Pillow
pandas
openpyxl